Fix a bug in one of the Android regexps.
This regexp:
   'Android[\- ][\d]+(?:\.[\d]+){1,2}; *[a-z]{0,2}[_\-]?[A-Za-z]{0,2};? Build'
lacks a capturing group (parens) for the device name.  This means that
it needs to specify the device-name specifically in the yaml file,
but it wasn't before.  This fixes that.

I also re-generated the json file from the yaml by running the second
half of convert.py.

Auditors: anastassia

Test Plan:
webapp% python
>>> import third_party.user_agents as u
>>> u.parse('Mozilla/5.0 (Linux; Android 5.1.99; Build/MRA23D) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/44.0.2403.106 Mobile Safari/537.36')

and it no longer raises an exception.  This user-agent was obtained
from bigquery like so:

  SELECT  user_agent
  from logs.requestlogs_20150731
  where  app_logs.message contains "no such group"
  group by user_agent